Some refactoring cleanup to keep our main file lighter weight.
